1- Langchain Basic flow/1.2 ollama/app.py
import os
from dotenv import load_dotenv
load_dotenv()

from langchain_community.llms import Ollama
import streamlit as st
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import StrOutputParser

## Langsmith Tracking
os.environ["LANGCHAIN_API_KEY"]=os.getenv("LANGCHAIN_API_KEY")
os.environ["LANGCHAIN_TRACING_V2"]="true"
os.environ["LANGCHAIN_PROJECT"]=os.getenv("LANGCHAIN_PROJECT")

# Prompt Template
prompt = ChatPromptTemplate.from_messages(
    [
        ("system", "You are the helpful assistance. Please respond to the question asked"),
        ("user", "Question:{question}")
    ]
)


# Streamlit framework
st.title("Langchain Demo with OLLAMA Llama 3.2 - 3B")
input_text = st.text_input("what question you have in mind?")


# OLLAMA Llama 3.2  model
llm = Ollama(model="llama3.2")
output_parser = StrOutputParser()
chain = prompt|llm|output_parser

if input_text:
    st.write(chain.invoke({"question":input_text}))