requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
wikipedia
pymupdf
fitz
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain-ollama
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
pydantic==1.10.12